+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Improvements:
 - Dictionary
   - Image search and DeepL supported on MacOS and Windows
   - Back navigation clears search and resets current entry
+  - Readability of kanji stroke order diagrams
 - Kana table: playback on android does not cut off
 - Anki: setting to decide how adding a duplicate should be handled